mmt-67 Removed AdminSessionManagement model and added view model.
