+++ app/code/Codifi/Training/Model/AdminSessionManagement.php
@@ -64,19 +64,6 @@ class AdminSessionManagement
     }
 
     /**
-     * Get credit_hold attribute value from current customer.
-     *
-     * @return bool
-     */
-    private function getCustomerAttrCreditHold(): bool
-    {
-        $customerData = $this->backendSession->getCustomerData();
-
-        return isset($customerData['account'][AddCustomerAttributeCreditHold::ATTRIBUTE_CODE]) &&
-            $customerData['account'][AddCustomerAttributeCreditHold::ATTRIBUTE_CODE];
-    }
-
-    /**
      * Get current customer id from admin session.
      *
      * @return int
@@ -108,58 +95,8 @@ class AdminSessionManagement
      *
      * @return array
      */
-    private function getCustomerIds(): array
+    public function getCustomerIds(): array
     {
         return $this->authSession->getData(self::ADMIN_SESSION_ATTRIBUTE_CUSTOMER_IDS) ?? [];
-    }
-
-    /**
-     * Check to be customer id in admin session array.
-     *
-     * @return bool
-     */
-    public function checkCustomerIdInArrayAdminSession(): bool
-    {
-        $currentCustomerId = $this->getCustomerId();
-        $customerIds = $this->getCustomerIds();
-        $isCustomerIdInArray = false;
-        if (!empty($customerIds) && in_array($currentCustomerId, $customerIds)) {
-            $isCustomerIdInArray = true;
-        }
-
-        return $isCustomerIdInArray;
-    }
-
-    /**
-     * Get options enabled.
-     *
-     * @return bool
-     */
-    public function isCreditHoldConfigEnabled(): bool
-    {
-        return $this->configProvider->isOptionCreditHoldEnable();
-    }
-
-    /**
-     * Last check for once show.
-     *
-     * @return bool
-     */
-    public function checkForOneTimeDemoMessage(): bool
-    {
-        return $this->getCustomerAttrCreditHold() &&
-            $this->isCreditHoldConfigEnabled() && !$this->checkCustomerIdInArrayAdminSession();
-    }
-
-    /**
-     * Get message and set customer id to admin session.
-     *
-     * @return string
-     */
-    public function getMessageAndSetCustomerIdToAdminSession(): string
-    {
-        $this->setCustomerIdToAdminSession();
-
-        return $this->configProvider->getMessage();
     }
 }
