+++ app/code/Codifi/Training/view/adminhtml/layout/customer_index_edit.xml
@@ -11,8 +11,12 @@
       xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
     <body>
         <referenceContainer name="after.body.start">
-            <block class="Codifi\Training\Block\Adminhtml\Edit\Tab\View\PersonalInfo"
-                   name="codifi.training.layout.personalinfo" template="Codifi_Training::tab/view/personal_info.phtml"/>
+            <block name="codifi.training.layout.personalinfo" template="Codifi_Training::tab/view/personal_info.phtml">
+                <arguments>
+                    <argument name="admin_session_management_view_model"
+                              xsi:type="object">Codifi\Training\ViewModel\Admin\AdminSessionManagement</argument>
+                </arguments>
+            </block>
         </referenceContainer>
     </body>
 </page>
